+++ cv_next/app/cv/[cvId]/components/commentForm/index.ts
@@ -1,0 +1,4 @@
+import { CommentForm } from "./commentForm";
+import { CommentFormRSC } from "./CommentFormRSC";
+
+export { CommentForm, CommentFormRSC };
