+++ cv_next/app/feed/components/cvItem/index.ts
@@ -1,0 +1,4 @@
+import { CVItem } from "./CVItem";
+import { CVItemRSC } from "./CVItemRSC";
+
+export { CVItem, CVItemRSC };
